bot/requirements.txt
aiogram==3.4.1
python-dotenv
django>=4.2
psycopg2-binary
openpyxl